ContentProvider Stuffs.
I've found that there was a typo in the Users table in the sqlite which
may have caused the initial problems with saving to the database but I'm
also getting a classnotfoundexception when trying to do an insert on the
contentresolver.

Probably just my misunderstanding as to how this is supposed to work but
I need another set of eyes on this and to figure out why that exception
is getting thrown deep within the call stack during the insert.